Create implementation of noughts and crosses game
The Game class performs three operations.

 1. Check that the specified grid is valid
 2. Check if there is a winner
 3. Make the next move if one is available

A valid grid is defined as:
 1. The number of Xs and Os are equal (including none of each)
 2. OR there is only one more X or O
 3. There is not more than one vertical or one horizontal winning move

A winning move is defined as three of the same value in a row either
horizontally, vertically, or diagonally in any direction. If there are
three in a row but the grid is invalid, no winner is reported.